requirements.txt
# Web server
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
jinja2==3.1.2
python-multipart==0.0.5

# Document processing
python-docx==0.8.11
PyPDF2==3.0.1
markdown==3.4.3
beautifulsoup4==4.12.2
requests==2.28.2
pandas==2.0.0
openpyxl==3.1.2

# Vector database
chromadb==0.4.6

# LLM
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.0+cu118
transformers==4.30.2
accelerate==0.19.0
bitsandbytes==0.38.0
sentence-transformers==2.2.2
